GeoIP module moved to core (#32)
